Move functions to services and utils
